🎨  abstract clickaway logic to a reusable hook (#347)
